# ~3x speedup, refactoring