… TWL06.txt
… TWL06Trimmed.txt
… TWL06Upper.txt
… prefixes7L.txt
… prefixes8L.txt