I managed to get the core board solving logic from 0.27s to 0.11 seconds by not adding words to words_info if they're not actual words - these are filtered out later and do not need to be added in the first place.
I also optimized a few other things, such as checking whether a word has a multiplier, and traversing the path.
I also added some pep8 formatting and code organization.