@philippeitis 2020-01-04
I've also converted TWL06.txt to TWL06Trimmed.txt - this file contains only words which are <= 12 characters and in their uppercase form - this saves 10% space, and it also allows skipping the .lower() operations.

I've also extended the prefixes to 8 characters, as I have found that this doubles the performance of the dfs:

```
2 to 02: 5.442s
2 to 03: 4.042s
2 to 04: 1.726s
2 to 05: 0.457s
2 to 06: 0.107s
2 to 07: 0.048s
2 to 08: 0.056s
2 to 09: 0.068s
2 to 10: 0.078s
2 to 11: 0.088s
2 to 12: 0.093s
```

(all tests include time to read the prefix files).

The speed up is now 11x for the DFS (0.27s to 0.0236s), and 3x overall, from 0.30s to 0.11s on my machine.