@philippeitis 2020-01-05
> > I've also converted TWL06.txt to TWL06Trimmed.txt - this file contains only words which are <= 12 characters and in their uppercase form - this saves 10% space, and it also allows skipping the .lower() operations.
> > I've also extended the prefixes to 8 characters, as I have found that this doubles the performance of the dfs:
> > ```
> > 2 to 02: 5.442s
> > 2 to 03: 4.042s
> > 2 to 04: 1.726s
> > 2 to 05: 0.457s
> > 2 to 06: 0.107s
> > 2 to 07: 0.048s
> > 2 to 08: 0.056s
> > 2 to 09: 0.068s
> > 2 to 10: 0.078s
> > 2 to 11: 0.088s
> > 2 to 12: 0.093s
> > ```
> > 
> > 
> > (all tests include time to read the prefix files).
> > The speed up is now 11x for the DFS (0.27s to 0.0236s), and 3x overall, from 0.30s to 0.11s on my machine.
> 
> My own testing done with timeit.timeit() and 100 iterations varies a lot, but I'm pretty sure the fastest max prefix length to check is 6-8. Your test seems to show a significant difference between 6 and 7 and not a significant difference between 7 and 8, so we can keep 8. How did you run your tests by the way?

The testing code looked like this. The main source of variance for me was file operations - it's usually a good idea to look at the lowest running time across all runs, as this is the run with the least actual overhead (eg. other processes using the disk). I modified the relevant functions to use LOWER_LIMIT and UPPER_LIMIT.

```
if __name__ == '__main__':
    # configs
    print_info = True  # if True, outputs information about board: number of words and total score
    from timeit import default_timer

    # directory with all necessary files
    main_dir = Path('./')

    # Reading data
    dictionary = get_dict()
    board, word_mults = read_board_from_file()
    word_int_mults = word_mults_to_int_array()
    points = get_points(board, word_mults)

    for i in range(2, 12):

        for j in range(3, 13):
            LOWER_LIMIT = i
            UPPER_LIMIT = j

            start = default_timer()
            prefixes = get_prefixes()

            # Solving
            words_info = []
            all_combos(board, points, word_int_mults)
            words_info = check_words(words_info)
            end = default_timer()
            print(f"{i} to {j}: {end - start} s")
```